+++ tos/admin.py
@@ -5,6 +5,8 @@ from tos.models import TermsOfService, UserAgreement
 
 class TermsOfServiceAdmin(admin.ModelAdmin):
     model = TermsOfService
+    list_display = ("slug", "created", "active")
+    list_filter = ("slug", "active")
 
 
 admin.site.register(TermsOfService, TermsOfServiceAdmin)